Updated and moved endpoints. Fixed all Rspecs.
