+++ lib/leankitkanban.rb
@@ -3,7 +3,7 @@ require "httparty"
 require "json"
 
 require File.join(File.dirname(__FILE__), 'leankit_request')
-require File.join(File.dirname(__FILE__), 'leankit_endpoint')
+require File.join(File.dirname(__FILE__), 'leankitkanban', 'endpoint')
 require File.join(File.dirname(__FILE__), 'leankitkanban', 'config')
 require File.join(File.dirname(__FILE__), 'leankitkanban', 'board')
 require File.join(File.dirname(__FILE__), 'leankitkanban', 'card')
